Issues understanding how Docker works
I am trying to package the HTML files into the 2 images, and post to my repo so that I can pull in and create ECS with the docker image. However I am not understanding how to then pull this back with docker compose. The files seem to copy to the image but then on a fresh directory I cannot have a running web server from the same image...